[Pluggable persistence] Make sure that ByteArrayRef is overrideable (some persistence impl will not use it)

+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/persistence/entity/JobEntityManagerImpl.java
@@ -300,8 +300,9 @@ public class JobEntityManagerImpl extends AbstractEntityManager<JobEntity> imple
    */
   protected void deleteExceptionByteArrayRef(JobEntity jobEntity) {
     ByteArrayRef exceptionByteArrayRef = jobEntity.getExceptionByteArrayRef();
-
-    exceptionByteArrayRef.delete();
+    if (exceptionByteArrayRef != null) {
+      exceptionByteArrayRef.delete();
+    }
   }
   
   // Job Execution logic ////////////////////////////////////////////////////////////////////
